moved database serializers to more logical place
