+++ mRemoteNGTests/Config/Serializers/DataTableDeserializerTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Data;
 using System.Linq;
 using mRemoteNG.Config.Serializers;
+using mRemoteNG.Config.Serializers.MsSql;
 using mRemoteNG.Connection;
 using mRemoteNG.Security;
 using mRemoteNG.Tree;
