+++ mRemoteV1/Config/Connections/SqlConnectionsSaver.cs
@@ -8,6 +8,7 @@ using mRemoteNG.App.Info;
 using mRemoteNG.Config.DatabaseConnectors;
 using mRemoteNG.Config.DataProviders;
 using mRemoteNG.Config.Serializers;
+using mRemoteNG.Config.Serializers.MsSql;
 using mRemoteNG.Config.Serializers.Versioning;
 using mRemoteNG.Container;
 using mRemoteNG.Messages;
